Moved LosslessStaking.js to environmentTests.js - Renamed utilsV3 to utils
